Create GameData
The GameData object can be thought of as the "football" in this software.  It is an instance of the game of RPG Grid Brawl at a particular moment in time.  It contains an array of Brawler objects and a 3D array of BoardSpace objects, among several other things.  This object gets passed in and out of many methods throughout the program.